% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Populations_hap.R
\name{create_haplo_pop_dataset}
\alias{create_haplo_pop_dataset}
\title{Calculate Haplotypes Differences between population}
\usage{
create_haplo_pop_dataset(filename)
}
\arguments{
\item{filename}{.arp file intended to be used for the analysis}
}
\value{
A text summary file and diagnostic plot gets saved in the current directory for the model.

A density plot is also saved in the environment to show the Haplotype differences per region.
}
\description{
This package builds a model to test if populations and regions can explain the differences in haplotypes sequences between individuals. You need to provide a .arp file with subject from different regions and their haplotypes sequence
}
\examples{
create_haplo_pop_dataset("example.arp")
create_haplo_pop_dataset("~/dir/example.arp")

}
